fr/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Pillow==10.0.0
pandas==2.1.3
face-recognition==1.3.0
